# docs/swagger.yaml
basePath: /
definitions:
  controllers.LoginRequest:
    properties:
      captcha:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captcha
    - password
    - username
    type: object
...
  controllers.emailConfigRequest:
    properties:
      enabled:
        type: boolean
      fromEmail:
        type: string
      fromName:
        type: string
      host:
        type: string
      password:
        type: string
      port:
        type: integer
      username:
        type: string
    required:
    - fromEmail
    - fromName
    - host
    - port
    - username
    type: object
...
  controllers.previewEmailTemplateRequest:
    properties:
      vars:
        additionalProperties:
          type: string
        type: object
    type: object
...
  controllers.updateEmailTemplateRequest:
    properties:
      content:
        type: string
      name:
        type: string
      subject:
        type: string
    required:
    - content
    - name
    - subject
    type: object
...
  models.EmailConfig:
    properties:
      enabled:
        type: boolean
      fromEmail:
        type: string
      fromName:
        type: string
      host:
        type: string
      id:
        type: integer
      port:
        type: integer
      username:
        type: string
    type: object
  models.EmailTemplate:
    properties:
      content:
        type: string
      createTime:
        type: string
      id:
        type: integer
      isBuiltin:
        type: boolean
      name:
        type: string
      scene:
        type: string
      subject:
        type: string
      updateTime:
        type: string
    type: object
  models.LoginLog:
    properties:
      createTime:
        type: string
      id:
        type: integer
      ip:
        type: string
      message:
        type: string
      success:
        type: boolean
      userAgent:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
...
  models.PageData:
    properties:
      pageData: {}
      total:
        type: integer
    type: object
...
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      originUrl:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: kite-admin 后台管理系统 API 文档
  title: Kite Admin API
  version: "1.0"
paths:
...
  /auth/current-role/switch/{roleCode}:
    post:
      description: 切换当前用户的角色，返回新的Token
      parameters:
      - description: 角色编码
        in: path
        name: roleCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功，data包含accessToken
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: 生成Token失败
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 切换当前角色
      tags:
      - 认证管理
  /auth/login:
    post:
      consumes:
      - application/json
      description: 通过用户名、密码和验证码登录系统
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功，data包含accessToken
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      summary: 用户登录
      tags:
      - 认证管理
  /auth/logout:
    post:
      description: 退出登录（客户端应清除Token）
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证管理
  /email-template/{id}:
    get:
      description: 根据ID获取邮件模板详情
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.EmailTemplate'
              type: object
        "404":
          description: 模板不存在
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 获取邮件模板详情
      tags:
      - 邮件模板
    put:
      consumes:
      - application/json
      description: 更新指定邮件模板的内容
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      - description: 模板信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.updateEmailTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.EmailTemplate'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: 模板不存在
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 更新邮件模板
      tags:
      - 邮件模板
  /email-template/{id}/preview:
    post:
      consumes:
      - application/json
      description: 使用变量渲染邮件模板并返回预览
      parameters:
      - description: 模板ID
        in: path
        name: id
        required: true
        type: integer
      - description: 模板变量
        in: body
        name: body
        schema:
          $ref: '#/definitions/controllers.previewEmailTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功，data包含subject和htmlBody
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: 模板不存在
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 预览邮件模板
      tags:
      - 邮件模板
  /email-template/list:
    get:
      description: 获取所有邮件模板
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.EmailTemplate'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取邮件模板列表
      tags:
      - 邮件模板
  /email/config:
    get:
      description: 获取当前邮件服务配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.EmailConfig'
              type: object
      security:
      - BearerAuth: []
      summary: 获取邮件配置
      tags:
      - 邮件配置
    put:
      consumes:
      - application/json
      description: 创建或更新邮件服务配置
      parameters:
      - description: 配置信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.emailConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.EmailConfig'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 保存邮件配置
      tags:
      - 邮件配置
  /email/config/test:
    post:
      description: 发送测试邮件到当前用户的邮箱
      produces:
      - application/json
      responses:
        "200":
          description: 测试邮件已加入队列
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 配置未启用或未绑定邮箱
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 测试邮件配置
      tags:
      - 邮件配置
  /loginlog/list:
    get:
      description: 分页查询登录成功日志，支持按用户名筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: pageNo
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 用户名（模糊搜索）
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/models.PageData'
                  - properties:
                      pageData:
                        items:
                          $ref: '#/definitions/models.LoginLog'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询登录日志
      tags:
      - 登录日志
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"